+++ src/MarketSignal.Contracts/Indicator/IndicatorRow.cs
@@ -4,5 +4,5 @@ namespace MarketSignal.Contracts.Indicator;
 
 public record IndicatorRow(
     Instant Time,
-    double Value
+    decimal Value
 );
