+++ src/MarketSignal.Contracts/Instrument/RawData/InstrumentRawDataRowEntity.cs
@@ -5,11 +5,11 @@ namespace MarketSignal.Contracts.RawData;
 public record InstrumentRawDataRowEntity(
     long Id,
     Instant Time,
-    double Open,
-    double High,
-    double Low,
-    double Close,
-    double Volume
+    decimal Open,
+    decimal High,
+    decimal Low,
+    decimal Close,
+    decimal Volume
 ) {
     public InstrumentRawDataRow ToDomain() => new(Time, Open, High, Low, Close, Volume);
 }
